// src/store/actions/action-types.ts
export enum ActionTypes {
    GET_BASIC_CONTENT = 'GET_BASIC_CONTENT',
    GET_BASIC_CONTENT_SUCCESS = 'GET_BASIC_CONTENT_SUCCESS',
    GET_BASIC_CONTENT_ERROR = 'GET_BASIC_CONTENT_ERROR',
    GET_MENUS = "GET_MENUS",
    GET_MENUS_SUCCESS = "GET_MENUS_SUCCESS",
    GET_MENUS_ERROR = "GET_MENUS_ERROR",
    GET_HOMEPAGE_CONTENT = "GET_HOMEPAGE_CONTENT",
    GET_HOMEPAGE_CONTENT_SUCCESS = "GET_HOMEPAGE_CONTENT_SUCCESS",
    GET_HOMEPAGE_CONTENT_ERROR = "GET_HOMEPAGE_CONTENT_ERROR",
    GET_ABOUTPAGE_CONTENT = "GET_ABOUTPAGE_CONTENT",
    GET_ABOUTPAGE_CONTENT_SUCCESS = "GET_ABOUTPAGE_CONTENT_SUCCESS",
    GET_ABOUTPAGE_CONTENT_ERROR = "GET_ABOUTPAGE_CONTENT_ERROR",
    COMPLETED_LAND_DATA_LOAD = "COMPLETED_LAND_DATA_LOAD",
    SET_SELECTED_TAB_BASIC_DETAILS = "SET_SELECTED_TAB_BASIC_DETAILS",
    GET_MENUS_DATA = "GET_MENUS_DATA",
    GET_MENUS_DATA_SUCCESS = "GET_MENUS_DATA_SUCCESS",
    GET_MENUS_DATA_ERROR = "GET_MENUS_DATA_ERROR",
    GET_ABOUT_US_PAGE_MENU_CONTENT = "GET_ABOUT_US_PAGE_MENU_CONTENT",
    GET_ABOUT_US_PAGE_MENU_CONTENT_SUCCESS = "GET_ABOUT_US_PAGE_MENU_CONTENT_SUCCESS",
    GET_ABOUT_US_PAGE_MENU_CONTENT_ERROR = "GET_ABOUT_US_PAGE_MENU_CONTENT_ERROR",
    SET_SELECTED_ABOUT_TAB = "SET_SELECTED_ABOUT_TAB",
    GET_OBJECT_PAGE_CONTENT = "GET_OBJECT_PAGE_CONTENT",
    GET_OBJECT_PAGE_CONTENT_SUCCESS = "GET_OBJECT_PAGE_CONTENT_SUCCESS",
    GET_OBJECT_PAGE_CONTENT_ERROR = "GET_OBJECT_PAGE_CONTENT_ERROR",
}